[refactor] TournamentMatchingInfoをコンポーネントとして分離
