+++ spa/views/TournamentMatching.js
@@ -3,40 +3,20 @@ import MatchingRoom, {
 } from "../components/MatchingRoom.js";
 import TitleMatchingRoom from "../components/TitleMatchingRoom.js";
 import WaitOrStart, { renderWaitOrStart } from "../components/WaitOrStart.js";
+import TournamentMatchingInfo, {
+  renderMatchingInfo,
+} from "../services/tournament/TournamentMatchingInfo.js";
 
 export default function TournamentMatching() {
-  function tournamentMatchingInfo() {
-    return `
-        <div id="tournament-matching-info" class="d-flex flex-column justify-content-center align-items-center text-center">
-          <span>
-            <span id="current-players">3</span>/<span id="max-players">16</span> players
-          </span>
-          <span>
-            start in <span id="matching-remain-sec">[59]</span> sec
-          </span>
-        </div>
-      `;
-  }
-
   return `
       ${TitleMatchingRoom("TOURNAMENT")}
-      ${tournamentMatchingInfo()}
+      ${TournamentMatchingInfo()}
       ${MatchingRoom()}
       ${WaitOrStart()}
 	`;
 }
 
 export function setupTournamentMatching() {
-  function changeMatchingInfo(playerSize, maxPlayerSize, remainSec) {
-    const currentPlayers = document.getElementById("current-players");
-    const maxPlayers = document.getElementById("max-players");
-    const matchingRemainSec = document.getElementById("matching-remain-sec");
-
-    currentPlayers.textContent = playerSize;
-    maxPlayers.textContent = maxPlayerSize;
-    matchingRemainSec.textContent = `[${remainSec}]`;
-  }
-
   const jsonData = [
     { avatarPath: "/assets/user.png", name: "rikeda1" },
     { avatarPath: "/assets/user.png", name: "rikeda2" },
@@ -45,6 +25,6 @@ export function setupTournamentMatching() {
   ];
 
   renderMatchingRoom(jsonData);
-  changeMatchingInfo(4, 16, 9);
+  renderMatchingInfo(4, 16, 9);
   renderWaitOrStart("Wait...", "#0ca5bf");
 }
